requirements.txt
Flask==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.1.1
email-validator==2.2.0
gunicorn==23.0.0
